+++ Samples/tvOS-Swift/tvOS-Swift/AppDelegate.swift
@@ -9,14 +9,14 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
-        SentrySDK.start(options: [
-                   "dsn": "https://8ee5199a90354faf995292b15c196d48@o19635.ingest.sentry.io/4394",
-                   "debug": true,
-                   "logLevel": "verbose",
-                   "enableAutoSessionTracking": true,
-                   "attachStacktrace": true,
-                   "sessionTrackingIntervalMillis": 5_000 
-               ])
+        SentrySDK.start { options in
+            options.dsn = "https://8ee5199a90354faf995292b15c196d48@o19635.ingest.sentry.io/4394"
+            options.debug = true
+            options.logLevel = SentryLogLevel.verbose
+            options.enableAutoSessionTracking = true
+            options.attachStacktrace = true
+            options.sessionTrackingIntervalMillis = 5_000
+        }
 
         // Create the SwiftUI view that provides the window contents.
         let contentView = ContentView()