+++ Sources/Sentry/SentryCrashIntegration.m
@@ -131,7 +131,19 @@ SentryCrashIntegration ()
             [outerScope setContextValue:appData forKey:@"app"];
 
             [outerScope addScopeListener:^(SentryScope *_Nonnull scope) {
-                [SentryCrash.sharedInstance setUserInfo:[scope serialize]];
+                NSMutableDictionary<NSString *, id> *userInfo =
+                    [[NSMutableDictionary alloc] initWithDictionary:[scope serialize]];
+
+                // SentryCrashReportConverter.convertReportToEvent needs the release name and the
+                // dist of the SentryOptions in the UserInfo. When SentryCrash records a crash it
+                // writes the UserInfo into SentryCrashField_User of the report.
+                // SentryCrashReportConverter.initWithReport loads the contents of
+                // SentryCrashField_User into self.userContext and convertReportToEvent can map the
+                // release name and dist to the SentryEvent. Fixes GH-581
+                userInfo[@"release"] = self.options.releaseName;
+                userInfo[@"dist"] = self.options.dist;
+
+                [SentryCrash.sharedInstance setUserInfo:userInfo];
             }];
         }];
     }