+++ Sources/Sentry/SentryOptions.m
@@ -14,11 +14,38 @@
     ];
 }
 
+- (instancetype)init
+{
+    if (self = [super init]) {
+        self.enabled = @NO;
+
+        // TODO: Remove this side effect
+        SentrySDK.logLevel = kSentryLogLevelError;
+        self.logLevel = kSentryLogLevelError;
+
+        self.debug = @NO;
+        self.maxBreadcrumbs = defaultMaxBreadcrumbs;
+        self.integrations = SentryOptions.defaultIntegrations;
+        self.sampleRate = @1;
+        self.enableAutoSessionTracking = @NO;
+        self.sessionTrackingIntervalMillis = [@30000 unsignedIntValue];
+        self.attachStacktrace = @NO;
+
+        // Set default release name
+        NSDictionary *infoDict = [[NSBundle mainBundle] infoDictionary];
+        if (nil != infoDict) {
+            self.releaseName =
+                [NSString stringWithFormat:@"%@@%@+%@", infoDict[@"CFBundleIdentifier"],
+                          infoDict[@"CFBundleShortVersionString"], infoDict[@"CFBundleVersion"]];
+        }
+    }
+    return self;
+}
+
 - (_Nullable instancetype)initWithDict:(NSDictionary<NSString *, id> *)options
                       didFailWithError:(NSError *_Nullable *_Nullable)error
 {
-    self = [super init];
-    if (self) {
+    if (self = [self init]) {
         [self validateOptions:options didFailWithError:error];
         if (nil != error && nil != *error) {
             [SentryLog
@@ -26,18 +53,23 @@
                       andLevel:kSentryLogLevelError];
             return nil;
         }
-
-        // If no user-defined release, use default.
-        if (nil == self.releaseName) {
-            NSDictionary *infoDict = [[NSBundle mainBundle] infoDictionary];
-            if (nil != infoDict) {
-                self.releaseName = [NSString
-                    stringWithFormat:@"%@@%@+%@", infoDict[@"CFBundleIdentifier"],
-                    infoDict[@"CFBundleShortVersionString"], infoDict[@"CFBundleVersion"]];
-            }
-        }
     }
     return self;
+}
+
+- (void)setDsn:(NSString *)dsn
+{
+    NSError *error = nil;
+    self.parsedDsn = [[SentryDsn alloc] initWithString:dsn didFailWithError:&error];
+
+    if (nil == error) {
+        _dsn = dsn;
+        self.enabled = @YES;
+    } else {
+        self.enabled = @NO;
+        NSString *errorMessage = [NSString stringWithFormat:@"Could not parse the DSN: %@", error];
+        [SentryLog logWithMessage:errorMessage andLevel:kSentryLogLevelError];
+    }
 }
 
 /**
@@ -47,27 +79,21 @@
 - (void)validateOptions:(NSDictionary<NSString *, id> *)options
        didFailWithError:(NSError *_Nullable *_Nullable)error
 {
-
-    if (nil != [options objectForKey:@"debug"]) {
-        self.debug = [NSNumber numberWithBool:[[options objectForKey:@"debug"] boolValue]];
-    } else {
-        self.debug = @NO;
+    if (nil != options[@"debug"]) {
+        self.debug = @([options[@"debug"] boolValue]);
     }
 
     if ([self.debug isEqual:@YES]) {
         // In other SDKs there's debug=true + diagnosticLevel where we can
         // control how chatty the SDK is. Ideally we'd support all the levels
         // here, and perhaps name it `diagnosticLevel` to align more.
-        if ([@"verbose" isEqual:[options objectForKey:@"logLevel"]]) {
+        if ([@"verbose" isEqual:options[@"logLevel"]]) {
             SentrySDK.logLevel = kSentryLogLevelVerbose;
             _logLevel = kSentryLogLevelVerbose;
         } else {
             SentrySDK.logLevel = kSentryLogLevelDebug;
             _logLevel = kSentryLogLevelDebug;
         }
-    } else {
-        SentrySDK.logLevel = kSentryLogLevelError;
-        _logLevel = kSentryLogLevelError;
     }
 
     if (nil == [options valueForKey:@"dsn"]
@@ -78,78 +104,62 @@
         return;
     }
 
-    self.dsn = [[SentryDsn alloc] initWithString:[options valueForKey:@"dsn"]
-                                didFailWithError:error];
+    self.parsedDsn = [[SentryDsn alloc] initWithString:[options valueForKey:@"dsn"]
+                                      didFailWithError:error];
     if (nil != error && nil != *error) {
         self.enabled = @NO;
     }
 
-    if ([[options objectForKey:@"release"] isKindOfClass:[NSString class]]) {
-        self.releaseName = [options objectForKey:@"release"];
+    if ([options[@"release"] isKindOfClass:[NSString class]]) {
+        self.releaseName = options[@"release"];
     }
 
-    if ([[options objectForKey:@"environment"] isKindOfClass:[NSString class]]) {
-        self.environment = [options objectForKey:@"environment"];
+    if ([options[@"environment"] isKindOfClass:[NSString class]]) {
+        self.environment = options[@"environment"];
     }
 
-    if ([[options objectForKey:@"dist"] isKindOfClass:[NSString class]]) {
-        self.dist = [options objectForKey:@"dist"];
+    if ([options[@"dist"] isKindOfClass:[NSString class]]) {
+        self.dist = options[@"dist"];
     }
 
-    if (nil != [options objectForKey:@"enabled"]) {
-        self.enabled = [NSNumber numberWithBool:[[options objectForKey:@"enabled"] boolValue]];
+    if (nil != options[@"enabled"]) {
+        self.enabled = @([options[@"enabled"] boolValue]);
     } else {
         self.enabled = @YES;
     }
 
-    if (nil != [options objectForKey:@"maxBreadcrumbs"]) {
-        self.maxBreadcrumbs = [[options objectForKey:@"maxBreadcrumbs"] unsignedIntValue];
-    } else {
-        // fallback value
-        self.maxBreadcrumbs = defaultMaxBreadcrumbs;
+    if (nil != options[@"maxBreadcrumbs"]) {
+        self.maxBreadcrumbs = [options[@"maxBreadcrumbs"] unsignedIntValue];
     }
 
-    if (nil != [options objectForKey:@"beforeSend"]) {
-        self.beforeSend = [options objectForKey:@"beforeSend"];
+    if (nil != options[@"beforeSend"]) {
+        self.beforeSend = options[@"beforeSend"];
     }
 
-    if (nil != [options objectForKey:@"beforeBreadcrumb"]) {
-        self.beforeBreadcrumb = [options objectForKey:@"beforeBreadcrumb"];
+    if (nil != options[@"beforeBreadcrumb"]) {
+        self.beforeBreadcrumb = options[@"beforeBreadcrumb"];
     }
 
-    if (nil != [options objectForKey:@"integrations"]) {
-        self.integrations = [options objectForKey:@"integrations"];
-    } else {
-        // fallback to defaultIntegrations
-        self.integrations = [SentryOptions defaultIntegrations];
+    if (nil != options[@"integrations"]) {
+        self.integrations = options[@"integrations"];
     }
 
-    NSNumber *sampleRate = [options objectForKey:@"sampleRate"];
-    if (nil == sampleRate || [sampleRate floatValue] < 0 || [sampleRate floatValue] > 1.0) {
-        self.sampleRate = @1;
-    } else {
+    NSNumber *sampleRate = options[@"sampleRate"];
+    if (nil != sampleRate && [sampleRate floatValue] >= 0 && [sampleRate floatValue] <= 1.0) {
         self.sampleRate = sampleRate;
     }
 
-    if (nil != [options objectForKey:@"enableAutoSessionTracking"]) {
-        self.enableAutoSessionTracking = [NSNumber
-            numberWithBool:[[options objectForKey:@"enableAutoSessionTracking"] boolValue]];
-    } else {
-        self.enableAutoSessionTracking = @NO; // TODO: Opt-out?
+    if (nil != options[@"enableAutoSessionTracking"]) {
+        self.enableAutoSessionTracking = @([options[@"enableAutoSessionTracking"] boolValue]);
     }
 
-    if (nil != [options objectForKey:@"sessionTrackingIntervalMillis"]) {
+    if (nil != options[@"sessionTrackingIntervalMillis"]) {
         self.sessionTrackingIntervalMillis =
-            [[options objectForKey:@"sessionTrackingIntervalMillis"] unsignedIntValue];
-    } else {
-        self.sessionTrackingIntervalMillis = [@30000 unsignedIntValue];
+            [options[@"sessionTrackingIntervalMillis"] unsignedIntValue];
     }
 
-    if (nil != [options objectForKey:@"attachStacktrace"]) {
-        self.attachStacktrace =
-            [NSNumber numberWithBool:[[options objectForKey:@"attachStacktrace"] boolValue]];
-    } else {
-        self.attachStacktrace = @NO;
+    if (nil != options[@"attachStacktrace"]) {
+        self.attachStacktrace = @([options[@"attachStacktrace"] boolValue]);
     }
 }