+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## unreleased
 
+- feat: Simplified SDK start #580
+- fix: Custom release name for crash reports #590
+
 ## 5.1.5
 
 - feat: Attach the stacktrace to custom events #583